% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_subs_ext.R
\name{file_subs_ext}
\alias{file_subs_ext}
\title{Replace file extension by another file extension}
\usage{
file_subs_ext(fn, new_ext, check = TRUE)
}
\arguments{
\item{fn}{character vector with file names}

\item{new_ext}{character vector of length one with the new extension (when
it does not start with a period a period is added).}

\item{check}{check if the new file name is not equal to the original
filename. If so, generate an error.}
}
\description{
Replace file extension by another file extension
}
