% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_eval.R
\name{format_traditional}
\alias{format_traditional}
\alias{format_copypaste}
\title{Format the result of running a block of code}
\usage{
format_traditional(x)

format_copypaste(x)
}
\arguments{
\item{x}{the result of running the code. See 'details' for the format.}
}
\value{
A character vector of length 1 with the formatted code.
}
\description{
Format the result of running a block of code
}
\details{
The input vector \code{x} should be a list. Each element of this list should
be a list with two elements: \code{input} and \code{output}. \code{input}
should contain the evaluated expression; this can be more than one line.
\code{output} should contain the output of the evaluation. When there is no
output this is character vector of length 0.
}
