% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extensions.R
\name{get_extensions}
\alias{get_extensions}
\title{Get the extensions that are needed when converting from json to markdown by
pandoc.}
\usage{
get_extensions()
}
\value{
Returns a character vector of the form \code{"+extension1-extension2"} with
the extensions that are to be enabled (\code{+}) or disables (\code{-}).
The function calls \code{pandoc} to check which extensions are available.
}
\description{
Get the extensions that are needed when converting from json to markdown by
pandoc.
}
\details{
Checks which extensions are available in pandoc and returns a character
string disabling extensions that interfere with the conversion of the
processed document (in json format) back to markdown. At the moment only one
extension is disabled.

\describe{
\item{\code{raw_attribute}}{ This extension is disabled. Some of the output
functions write raw markdown blocks. This extension only recognises html
and tex raw blocks. Raw markdown blocks are ignored when converting to
other formats. }
}
}
