% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_block.R
\name{markdown_block}
\alias{markdown_block}
\title{Return a code block object that can be included in the pandoc parse tree}
\usage{
markdown_block(content, language, id = "", ...)
}
\arguments{
\item{content}{a character vector containing the code}

\item{language}{language of the code in the code block}

\item{id}{optional id of the code block}

\item{...}{ignored.}
}
\description{
Return a code block object that can be included in the pandoc parse tree
}
