% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_table.R
\name{md_table}
\alias{md_table}
\title{Generate a markdown table from a data.frame}
\usage{
md_table(tab, caption, as_character = FALSE, ...)
}
\arguments{
\item{tab}{a data frame}

\item{caption}{text of the caption. When omitted no caption is added to the
table.}

\item{as_character}{return the table as a character vector. If \code{FALSE}
the table will be written to the standard output.}

\item{...}{unused.}
}
\value{
Then \code{as_character = FALSE} a character vector with the markdown
containing the table is returned.  Otherwise,
nothing is returned; the markdown is then written to the console.
}
\description{
Generate a markdown table from a data.frame
}
