% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_block.R
\name{raw_block}
\alias{raw_block}
\title{Return a raw object that can be included in the pandoc parse tree}
\usage{
raw_block(content, language = "markdown")
}
\arguments{
\item{content}{a character vector containing the content to include
in the final document.}

\item{language}{language of the content}
}
\description{
Return a raw object that can be included in the pandoc parse tree
}
\details{
A raw block is included as is into the final markdown document. This can
be used for example to include raw chunks of markdown.
}
